HTBHF-2571 Create v2 and v3 directories plus move correct implementations into those directories.
