+++ api/src/main/java/uk/gov/dhsc/htbhf/claimant/service/v2/EligibilityClient.java
@@ -1,4 +1,4 @@
-package uk.gov.dhsc.htbhf.claimant.service.v1;
+package uk.gov.dhsc.htbhf.claimant.service.v2;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
