+++ api/src/main/java/uk/gov/dhsc/htbhf/claimant/service/v2/IdentityAndEligibilityResponseFactory.java
@@ -1,4 +1,4 @@
-package uk.gov.dhsc.htbhf.claimant.service.v1;
+package uk.gov.dhsc.htbhf.claimant.service.v2;
 
 import org.springframework.util.CollectionUtils;
 import uk.gov.dhsc.htbhf.claimant.model.eligibility.EligibilityResponse;
