+++ api/src/main/java/uk/gov/dhsc/htbhf/claimant/service/v3/EligibilityClientV2.java
@@ -1,4 +1,4 @@
-package uk.gov.dhsc.htbhf.claimant.service.v2;
+package uk.gov.dhsc.htbhf.claimant.service.v3;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
