+++ api/src/test/java/uk/gov/dhsc/htbhf/claimant/controller/v2/ClaimControllerTest.java
@@ -1,4 +1,4 @@
-package uk.gov.dhsc.htbhf.claimant.controller;
+package uk.gov.dhsc.htbhf.claimant.controller.v2;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
