+++ api/src/test/java/uk/gov/dhsc/htbhf/claimant/service/v2/EligibilityAndEntitlementServiceV1Test.java
@@ -1,4 +1,4 @@
-package uk.gov.dhsc.htbhf.claimant.service.v1;
+package uk.gov.dhsc.htbhf.claimant.service.v2;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
