+++ api/src/test/java/uk/gov/dhsc/htbhf/claimant/service/v3/EligibilityClientV2Test.java
@@ -1,4 +1,4 @@
-package uk.gov.dhsc.htbhf.claimant.service.v2;
+package uk.gov.dhsc.htbhf.claimant.service.v3;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
