+++ api/src/test/java/uk/gov/dhsc/htbhf/claimant/service/v3/EligibilityOutcomeToEligibilityStatusConverterTest.java
@@ -1,4 +1,4 @@
-package uk.gov.dhsc.htbhf.claimant.service.v2;
+package uk.gov.dhsc.htbhf.claimant.service.v3;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
@@ -6,7 +6,7 @@ import uk.gov.dhsc.htbhf.dwp.model.v2.EligibilityOutcome;
 import uk.gov.dhsc.htbhf.eligibility.model.EligibilityStatus;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static uk.gov.dhsc.htbhf.claimant.service.v2.EligibilityOutcomeToEligibilityStatusConverter.fromEligibilityOutcome;
+import static uk.gov.dhsc.htbhf.claimant.service.v3.EligibilityOutcomeToEligibilityStatusConverter.fromEligibilityOutcome;
 
 class EligibilityOutcomeToEligibilityStatusConverterTest {
